+++ genti-external/src/main/java/com/gt/genti/discord/event/SignUpEvent.java
@@ -1,4 +1,4 @@
-package com.gt.genti.discord;
+package com.gt.genti.discord.event;
 
 import java.time.LocalDateTime;
