+++ genti-external/src/main/java/com/gt/genti/discord/event/SignUpEventListener.java
@@ -1,8 +1,10 @@
-package com.gt.genti.discord;
+package com.gt.genti.discord.event;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
+
+import com.gt.genti.discord.DiscordAppender;
 
 import lombok.RequiredArgsConstructor;
